move proposal threading methods to a Concern
